Setting resource name to fqn
